+++ test/done-success-test.lua
@@ -1,5 +1,5 @@
 package.path = '../src/?.lua;' .. package.path
-tap = require 'tapered'
+local tap = require 'tapered'
 
 tap.ok(true)
 tap.done(1)